+++ Sources/Module/Channel/GroupChannel/SBUGroupChannelModule.List.swift
@@ -150,7 +150,7 @@ extension SBUGroupChannelModule {
         }
         
         // MARK: Private properties
-        var voicePlayer: SBUVoicePlayer?
+        public var voicePlayer: SBUVoicePlayer?
         var voiceFileInfos: [String: SBUVoiceFileInfo] = [:]
         var currentVoiceFileInfo: SBUVoiceFileInfo?
         var currentVoiceContentView: SBUVoiceContentView?