+++ Sources/Module/ChannelList/SBUBaseChannelListModule.Header.swift
@@ -49,6 +49,7 @@ extension SBUBaseChannelListModule {
         // MARK: - UI properties (Public)
         
         /// A view that represents a title in navigation bar.
+        /// The default view type is ``SBUNavigationTitleView``.
         /// - NOTE: When the value is updated, `channelListModule(_:didUpdateTitleView:)` delegate function is called.
         public var titleView: UIView? {
             didSet { self.didUpdateTitleView() }