+++ Sources/Module/SelectUser/SBUBaseSelectUserModule.List.swift
@@ -41,6 +41,7 @@ extension SBUBaseSelectUserModule {
         public var tableView = UITableView()
         
         /// A view that shows when there is no message in the channel.
+        /// The default view type is ``SBUEmptyView``.
         public var emptyView: UIView? {
             didSet { self.tableView.backgroundView = self.emptyView }
         }